+++ components/Slides/Containers/Slides/FrostPane/index.tsx
